Pronoun resolution and Appositives
Still needs a little tweaking. Number agreement is sketchy, and gender
agreement only checks for modern English names. This should be extended
to handle arguments such as: Mr. Ms. Miss, etc.
Number agreement needs to simply check the context of the anaphora, and
the NP in question, for plurals.

Pronoun returns the potential pronoun in the form of a string.